software/timesync/requirements.txt
ntplib==0.4.0
pyserial==3.5
